string fields should no longer silently convert zeros to the string "0". it's still not optimal, as getting a field previously set using zero will return "".
however, with that exception in mind, export and import mostly works now.

the technique is still pretty ugly, as we could do with a dependency calculation algorithm.
also, many-to-many relationships now have the concept of "ownership", that is, when exporting/importing,
the m-2-m relationship data will be exported on both sides of the relationship, but will be imported only
on the "owning" side, which is defined by the first part of the m-2-m relationship table (for example,
in a group <-> user relationship, if the mapping table is called "users_groups", then the relationship will
be owned by the user type. note that this is important because there may be cyclical relationships in one direction
or the other. from a taxonomy standpoint, this example would be better expressed with the group having
ownership, thus groups_users.

ownership is a flag on the multikeyreference type, if you're creating the relationship manually.

also, read-only field types will not be imported/exported, as determined by the is_read_only flag on the field.