macros may have content now. syntax:
<%mymacro ...%>
content
<%/mymacro%>

macro will receive content within as the argument "contents". this will
be a macro renderer, not the string, so it must not be accessed as one.
this is to enable deferring the render of the contents unless/until it's
actually used.

in other words: it may be output as any other variable, though it may
not be tested as one.